Adicionado simples inteligência artificial para os inimigos utilizando o método Enemy.Update(). O método isFree() para teste de colisão foi movido para uma nova Class chamada Collision. Foi adicionado o método Collision.simulate() para prevenir com que o suposto tamanho do Array, obtido através do calculo especificado, seja maior do que o próprio Array. O método Collision.simulatePosition() está em testes e desenvolvimento, não se sabe ao certo se será algo necessário.

+++ src/com/zelda/entity/Avatar.java
@@ -30,13 +30,13 @@ public class Avatar {
 	}
 	
 	public void Render(Graphics graphics) {
-		graphics.drawImage(getSprite(), (int)getX() - Camera.x, (int)getY() - Camera.y, null);
+		graphics.drawImage(getSprite(), getX() - Camera.x, getY() - Camera.y, null);
 	}
-	
+
 	private void setX(int x) {
 		this.x = x;
 	}
-	
+
 	private void setY(int y) {
 		this.y = y;
 	}
